Extract `loadBenchmarkInfo` function
