+++ site/frontend/src/api.ts
@@ -1,24 +1,13 @@
-export async function postRequest<T>(path: string, body: any): Promise<T> {
-    const response = await fetch(path, {
-        method: "POST",
-        body: JSON.stringify(body),
-    });
-    return await response.json();
+import {getRequest} from "./utils/requests";
+import {INFO_URL} from "./urls";
+
+export interface BenchmarkInfo {
+  // Known statistic values from the DB
+  stats: [string];
+  // Last loaded run date
+  as_of: string | null;
 }
 
-export async function getRequest<T>(path: string, params: Dict<string | null> = {}): Promise<T> {
-    let url = path;
-
-    if (Object.keys(params).length > 0) {
-        const urlParams = new URLSearchParams();
-        for (const [key, value] of Object.entries(params)) {
-            if (value !== null) {
-                urlParams.set(key, value);
-            }
-        }
-        url = `${path}?${urlParams}`;
-    }
-
-    const response = await fetch(url, {});
-    return await response.json();
+export async function loadBenchmarkInfo() {
+  return await getRequest<BenchmarkInfo>(INFO_URL);
 }
